Refactor SongSettingsTab into column components
Extract SongSettingsTab's large column implementations into three new files (SongSettingsLeftColumn.kt, SongSettingsRightColumn.kt, SongSettingsLookAheadColumn.kt) and simplify SongSettingsTab to compose those columns. Make segmentedItemShape internal for reuse, restore the titleSlideShowSongNumber checkbox callback, and remove many now-unneeded imports/inline code. UI behaviour kept intact (auto-fit, presenter-aware controls remain in the Right column).
